# backend/modules/accounts/src/main/java/com/yourorg/quickapp/accounts/RegisterRequest.java
package com.yourorg.quickapp.accounts;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8, max = 128) String password,
        Boolean rememberMe) {}